pyproject.toml
[tool.poetry]
name = "hamal"
version = "0.1.0"
description = ""
authors = []
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
Django = "^5.1"
django-configurations = "^2.5.1"
dj-database-url = "^2.3.0"
djangorestframework = "^3.15.2"
django-filter = "^25.1"
django-extensions = "^3.2.3"
django-import-export = {extras = ["xlsx"], version = "^4.3.9"}
django-storages = "^1.14.5"
Pillow = "^11.1.0"
django-cors-headers = "^4.7.0"
django-querycount = "^0.8.3"
gunicorn = "^23.0.0"
djangorestframework-simplejwt = "^5.4.0"
psycopg = "^3.2.5"
drf-nested-routers = "^0.94.1"
drf-spectacular = "^0.28.0"
boto3 = ">=1.37.5"
django-simple-history = "^3.10.1"
django-admin-autocomplete-filter = "^0.7.1"
django-modeladmin-reorder = {git = "https://github.com/bloodwithmilk25/django-modeladmin-reorder.git", rev = "master"}






[tool.poetry.dev-dependencies]

pytest = "^8.3.5"
ruff = "^0.9.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88

# To never enforce `E501` (line length violations) uncomment the next line
# ignore = ["E501"]
